+++ CHANGELOG.md
@@ -8,6 +8,8 @@ To get the diff for a specific change, go to https://github.com/superdesk/web-pu
 
 * 1.3.0
  * feature [#459] Added an option to preview an article before even publishing it
+ * feature [#462] Added template widget
+ * bug [#460] Install assets when theme is installed
  * feature [#458] Added redirect and notFound functions for twig templates
  * feature [#454] Override existing rules by publish destinations
  * improvement [#455] Added slug to article author - allow to load authors by slug